Split sign up view into pages
