+++ lib/screens/welcome/views/_views.dart
@@ -1,1 +1,2 @@
 export 'login_view.dart';
+export 'sign_up_view/sign_up_view.dart';
