+++ lib/screens/welcome/views/sign_up_view/pages/_pages.dart
@@ -1,0 +1,3 @@
+export 'account_page.dart';
+export 'phone_page.dart';
+export 'verification_page.dart';
