+++ lib/screens/welcome/views/sign_up_view/pages/verification_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class VerificationPage extends StatefulWidget {
+  const VerificationPage({Key? key}) : super(key: key);
+
+  @override
+  _VerificationPageState createState() => _VerificationPageState();
+}
+
+class _VerificationPageState extends State<VerificationPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
